+++ tests/test_calculate_recurrent_infections.py
@@ -19,6 +19,7 @@ def test_recurrent_contact_infects_susceptibles_and_leaves_other_group_untouched
     infection_probabilities = np.array([1])
     infected = np.array([False, False, False, False])
     infection_counter = np.zeros(4, dtype=np.int_)
+    infection_probability_multiplier = np.ones(len(recurrent_contacts))
 
     (
         infected,
@@ -32,6 +33,7 @@ def test_recurrent_contact_infects_susceptibles_and_leaves_other_group_untouched
         group_codes,
         indexers,
         infection_probabilities,
+        infection_probability_multiplier,
         infected,
         infection_counter,
         0,
@@ -57,8 +59,9 @@ def test_infections_occur_not_in_other_recurrent_group():
     indexers.append(sub_indexers)
 
     infection_probabilities = np.array([1])
-    infected = np.array([False, True, False, False])
+    infected = np.array([False, False, False, False])
     infection_counter = np.zeros(4, dtype=np.int_)
+    infection_probability_multiplier = np.ones(len(recurrent_contacts))
 
     (
         infected,
@@ -72,6 +75,7 @@ def test_infections_occur_not_in_other_recurrent_group():
         group_codes,
         indexers,
         infection_probabilities,
+        infection_probability_multiplier,
         infected,
         infection_counter,
         0,
@@ -81,3 +85,45 @@ def test_infections_occur_not_in_other_recurrent_group():
     assert (infection_counter == [1, 0, 0, 0]).all()
     assert (immune == [True, True, False, False]).all()
     assert (was_infected == [-1, 0, -1, -1]).all()
+
+
+def test_infections_can_be_scaled_with_multiplier():
+    """Run a scenario where infections are halved by the multiplier."""
+    n_individuals = 100_000
+    recurrent_contacts = np.full((n_individuals, 1), True)
+    infectious = np.array([True] + [False] * (n_individuals - 1))
+    immune = np.array([True] + [False] * (n_individuals - 1))
+
+    group_codes = np.zeros((n_individuals, 1), dtype=np.int_)
+
+    sub_indexers = nb.typed.List()
+    sub_indexers.append(np.arange(n_individuals, dtype=np.int_))
+    indexers = nb.typed.List()
+    indexers.append(sub_indexers)
+
+    infection_probabilities = np.array([1])
+    infected = np.full(n_individuals, False)
+    infection_counter = np.zeros(n_individuals, dtype=np.int_)
+    infection_probability_multiplier = np.full(n_individuals, 0.5)
+
+    (
+        infected,
+        infection_counter,
+        immune,
+        was_infected,
+    ) = _calculate_infections_by_recurrent_contacts(
+        recurrent_contacts,
+        infectious,
+        immune,
+        group_codes,
+        indexers,
+        infection_probabilities,
+        infection_probability_multiplier,
+        infected,
+        infection_counter,
+        0,
+    )
+
+    assert np.isclose(infected.sum(), n_individuals / 2, atol=1e2)
+    assert np.isclose(infection_counter[0], n_individuals / 2, atol=1e2)
+    assert np.isclose(immune.sum(), n_individuals / 2, atol=1e2)